\name{mwe-travis-rmarkdown-vignette-package}
\alias{mwe-travis-rmarkdown-vignette-package}
\alias{mwe-travis-rmarkdown-vignette}
\docType{package}
\title{\packageTitle{mwe-travis-rmarkdown-vignette}}
\description{\packageDescription{mwe-travis-rmarkdown-vignette}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{mwe-travis-rmarkdown-vignette}
  \packageIndices{mwe-travis-rmarkdown-vignette}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{mwe-travis-rmarkdown-vignette}
  
  Maintainer: \packageMaintainer{mwe-travis-rmarkdown-vignette}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
